Enhance systemd support and document it
Since we encourage throw-away containers, make the systemd unit operate
in that manner.  Make use of --rm so the container is removed on exit.
Use --name %N (which expands to the new of the unit with the suffix
dropped) for the name.  Add --cap-add=NET_ADMIN as that is required and
pass in --dns 127.0.0.1 as it's required and --dns 8.8.8.8 for the
second DNS as this is the project-default public DNS server mentioned.
Make use of --net host for ease of integration.

Signed-off-by: Tom Rini <trini@konsulko.com>