Update install.sh
I was wondering why the pi-hole image was over 300MB for me and started taking a look. 

Noticed that install.sh is cloning full repos which seems unnecessary. The proposed change matches the clone depth of 10 that fetch_release_metadata() uses.

Doing so reduces the size of the repos from 45.7MB to 9.5MB. Not huge, I suppose but it's a start.

full repo
38M	adminLTE
7.7M	pi-hole

depth 10
8.1M	adminLTE
1.4M	pi-hole