Set DNSMASQ_LISTENING=all in the Dockerfile
This option controls the permit all origins flag[0] - setting this to local only
does not make logical sense for a container image build. While the web UI can
be used to toggle this option, one of the strengths of running pihole in a
container is that users can use it in a fully automated, toil free setup, which
this option impedes. The local-subnets option may be very useful when run in a
container either (depending on the definition of a local subnet) so I'm
proposing that we change it to `all`.


[0] https://docs.pi-hole.net/ftldns/interfaces/#all

Signed-off-by: alex <alex@stty.io>