as mentioned in https://github.com/pi-hole/docker-pi-hole/issues/1323
It is worth noting that this page will exist on a bare metal fresh installation, too. Should we delete it? Probably not - we shouldn't touch that directory really - but in docker we have more control as we build the image from fresh each time. TL;DR, we can do things here that we shouldn't in bare metal installs

Signed-off-by: Adam Warner <me@adamwarner.co.uk>