Allow custom web interface URI for Pi-hole admin
This change introduces the ability to customize the URI
from which the Pi-hole admin interface is served. Previously,
the webroot was hardcoded as "/admin/". Now, if the configuration
variable FTLCONF_webserver_paths_webhome is defined
(e.g., FTLCONF_webserver_paths_webhome=/pihole/), the interface
will be served from that specified path.

This update enhances flexibility for deployments that use a reverse proxy
to handle HTTPS and route traffic to the correct container path.

Signed-off-by: pepramon <pepramon@gmail.com>