Supplement docker secrets usage information
Docker secrets has a couple of characteristics that may cause baffling
errors where configuration looks correct but the pihole container cannot
read the secret. The UID and GID of the secret file on the host must be
set to the same value used by the pihole process in the container
(typically 1000).

Signed-off-by: Bradley G Smith <bradley.g.smith@gmail.com>