fix and improve the waiting until FTL has started
When the container is restarted, the FTL.log already contains the string
we are waiting for. In that case the check does not work as intended.

To fix that, the FTL.log lines are counted before FTL starts. `tail` then uses
that value (+1) to start from.

The loop with the sleep command has been replaced with a more elegant solution
I've found: https://superuser.com/a/900134

Signed-off-by: casperklein <casperklein@users.noreply.github.com>